Add jquery, and javascript asset pipeline structure for materializecss
When Rails new is run it doesn't include the stock asset pipeline
file structure that it used to come with. It now includes webpack and
uses a seperate file structure to compile all your javascript code.

I added the files needed and dependcies to solve create the pipeline.
I also added back in the jquery used by materialize to get the dropdown
menu to function correctly.